test : 랭킹 관련 API는 ranking 디렉토리로 이동 및 분리
