+++ src/test/java/com/gamzabat/algohub/service/StudyGroupServiceTest.java
@@ -26,9 +26,6 @@ import org.springframework.mock.web.MockMultipartFile;
 import com.gamzabat.algohub.enums.Role;
 import com.gamzabat.algohub.exception.StudyGroupValidationException;
 import com.gamzabat.algohub.exception.UserValidationException;
-import com.gamzabat.algohub.feature.group.ranking.domain.Ranking;
-import com.gamzabat.algohub.feature.group.ranking.dto.GetRankingResponse;
-import com.gamzabat.algohub.feature.group.ranking.exception.CannotFoundRankingException;
 import com.gamzabat.algohub.feature.group.ranking.repository.RankingRepository;
 import com.gamzabat.algohub.feature.group.studygroup.domain.BookmarkedStudyGroup;
 import com.gamzabat.algohub.feature.group.studygroup.domain.GroupMember;
@@ -74,14 +71,13 @@ class StudyGroupServiceTest {
 	private RankingRepository rankingRepository;
 	@Mock
 	private ImageService imageService;
-	private User user, owner, user2, user3, user4;
+	private User user, owner, user2, user3;
 	private StudyGroup group;
 	private Problem problem1, problem2;
 	private Solution solution1, solution2, solution3;
+	private GroupMember groupMember1, groupMember2, groupMember3;
 	private final Long groupId = 10L;
 	private GroupMember ownerGroupmember;
-	private GroupMember groupMember1, groupMember2, groupMember3, groupMember4;
-	private Ranking ranking1, ranking2, ranking3, ranking4;
 	@Captor
 	private ArgumentCaptor<StudyGroup> groupCaptor;
 	@Captor
@@ -97,8 +93,7 @@ class StudyGroupServiceTest {
 			.role(Role.USER).profileImage("image2").build();
 		user3 = User.builder().email("email3").password("password").nickname("nickname3")
 			.role(Role.USER).profileImage("image3").build();
-		user4 = User.builder().email("email4").password("password").nickname("nickname4")
-			.role(Role.USER).profileImage("image4").build();
+
 		group = StudyGroup.builder()
 			.name("name")
 			.startDate(LocalDate.now())
@@ -130,12 +125,6 @@ class StudyGroupServiceTest {
 			.role(RoleOfGroupMember.ADMIN)
 			.joinDate(LocalDate.now())
 			.build();
-		groupMember4 = GroupMember.builder()
-			.studyGroup(group)
-			.user(user4)
-			.role(RoleOfGroupMember.PARTICIPANT)
-			.joinDate(LocalDate.now())
-			.build();
 
 		problem1 = Problem.builder()
 			.studyGroup(group)
@@ -160,31 +149,6 @@ class StudyGroupServiceTest {
 			.solvedDateTime(LocalDateTime.now().minusDays(1))
 			.problem(problem1)
 			.user(user2)
-			.build();
-
-		ranking1 = Ranking.builder()
-			.member(groupMember1)
-			.solvedCount(3)
-			.currentRank(1)
-			.rankDiff("-")
-			.build();
-		ranking2 = Ranking.builder()
-			.member(groupMember2)
-			.solvedCount(2)
-			.currentRank(2)
-			.rankDiff("+1")
-			.build();
-		ranking3 = Ranking.builder()
-			.member(groupMember3)
-			.solvedCount(1)
-			.currentRank(3)
-			.rankDiff("-1")
-			.build();
-		ranking4 = Ranking.builder()
-			.member(groupMember4)
-			.solvedCount(0)
-			.currentRank(4)
-			.rankDiff("-")
 			.build();
 
 		Field userField = User.class.getDeclaredField("id");
@@ -571,144 +535,6 @@ class StudyGroupServiceTest {
 	}
 
 	@Test
-	@DisplayName("전체랭킹 조회 성공")
-	void getAllRank_SuccessByOwner() {
-		//given
-		when(studyGroupRepository.findById(10L)).thenReturn(Optional.of(group));
-		when(groupMemberRepository.existsByUserAndStudyGroup(user2, group)).thenReturn(true);
-		List<Ranking> ranking = new ArrayList<>();
-		ranking.add(ranking2);
-		ranking.add(ranking3);
-		ranking.add(ranking1);
-		ranking.add(ranking4);
-		when(rankingRepository.findAllByStudyGroup(group)).thenReturn(ranking);
-
-		//when
-		List<GetRankingResponse> result = studyGroupService.getAllRank(user2, 10L);
-
-		//then
-		assertThat(result.size()).isEqualTo(4);
-
-		for (int i = 0; i < 4; i++) {
-			assertThat(result.get(i).getProfileImage()).isEqualTo("image" + (i + 1));
-			assertThat(result.get(i).getSolvedCount()).isEqualTo(3 - i);
-			assertThat(result.get(i).getUserNickname()).isEqualTo("nickname" + (i + 1));
-			assertThat(result.get(i).getRank()).isEqualTo(i + 1);
-		}
-	}
-
-	@Test
-	@DisplayName("전체랭킹 조회 실패 : 그룹을 못 찾은 경우")
-	void getAllRank_FailedByCannotFoundGroup() {
-		//given
-		when(studyGroupRepository.findById(9L)).thenReturn(Optional.empty());
-
-		//then
-		assertThatThrownBy(() -> studyGroupService.getAllRank(user, 9L))
-			.isInstanceOf(CannotFoundGroupException.class)
-			.hasFieldOrPropertyWithValue("errors", "그룹을 찾을 수 없습니다.");
-	}
-
-	@Test
-	@DisplayName("전체랭킹 조회 실패 : 랭킹을 확인할 권한이 없는 경우")
-	void getAllRank_FailedByAccess() {
-		//given
-		when(studyGroupRepository.findById(groupId)).thenReturn(Optional.of(group));
-		when(groupMemberRepository.existsByUserAndStudyGroup(user2, group)).thenReturn(false);
-
-		//then
-		assertThatThrownBy(() -> studyGroupService.getAllRank(user2, groupId))
-			.isInstanceOf(GroupMemberValidationException.class)
-			.hasFieldOrPropertyWithValue("code", HttpStatus.FORBIDDEN.value())
-			.hasFieldOrPropertyWithValue("error", "랭킹을 확인할 권한이 없습니다.");
-	}
-
-	@Test
-	@DisplayName("Top 3 랭킹 조회 성공")
-	void getTopRanking() {
-		//given
-		when(studyGroupRepository.findById(10L)).thenReturn(Optional.of(group));
-		when(groupMemberRepository.existsByUserAndStudyGroup(user2, group)).thenReturn(true);
-		List<Ranking> ranking = new ArrayList<>();
-		ranking.add(ranking2);
-		ranking.add(ranking3);
-		ranking.add(ranking4);
-		ranking.add(ranking1);
-		when(rankingRepository.findAllByStudyGroup(group)).thenReturn(ranking);
-
-		//when
-		List<GetRankingResponse> result = studyGroupService.getTopRank(user2, 10L);
-
-		//then
-		assertThat(result.size()).isEqualTo(3);
-
-		for (int i = 0; i < 3; i++) {
-			assertThat(result.get(i).getProfileImage()).isEqualTo("image" + (i + 1));
-			assertThat(result.get(i).getSolvedCount()).isEqualTo(3 - i);
-			assertThat(result.get(i).getUserNickname()).isEqualTo("nickname" + (i + 1));
-			assertThat(result.get(i).getRank()).isEqualTo(i + 1);
-		}
-	}
-
-	@Test
-	@DisplayName("Top 3 랭킹 조회 성공 : 모두 0 solved인 경우")
-	void getTopRanking_EmptyRank() {
-		//given
-		when(studyGroupRepository.findById(10L)).thenReturn(Optional.of(group));
-		when(groupMemberRepository.existsByUserAndStudyGroup(user2, group)).thenReturn(true);
-		List<Ranking> ranking = new ArrayList<>();
-		ranking.add(Ranking.builder()
-			.member(groupMember1)
-			.solvedCount(0)
-			.build());
-		ranking.add(Ranking.builder()
-			.member(groupMember2)
-			.solvedCount(0)
-			.build());
-		ranking.add(Ranking.builder()
-			.member(groupMember3)
-			.solvedCount(0)
-			.build());
-		ranking.add(Ranking.builder()
-			.member(groupMember4)
-			.solvedCount(0)
-			.build());
-		when(rankingRepository.findAllByStudyGroup(group)).thenReturn(ranking);
-
-		//when
-		List<GetRankingResponse> result = studyGroupService.getTopRank(user2, 10L);
-
-		//then
-		assertThat(result.size()).isEqualTo(0);
-	}
-
-	@Test
-	@DisplayName("Top 3 랭킹 조회 실패 : 그룹을 못 찾은 경우")
-	void getTopRanking_FailedByCannotFoundGroup() {
-		//given
-		when(studyGroupRepository.findById(9L)).thenReturn(Optional.empty());
-
-		//then
-		assertThatThrownBy(() -> studyGroupService.getTopRank(user, 9L))
-			.isInstanceOf(CannotFoundGroupException.class)
-			.hasFieldOrPropertyWithValue("errors", "그룹을 찾을 수 없습니다.");
-	}
-
-	@Test
-	@DisplayName("Top 3 랭킹 조회 실패 : 랭킹을 확인할 권한이 없는 경우")
-	void getTopRanking_FailedByAccess() {
-		//given
-		when(studyGroupRepository.findById(groupId)).thenReturn(Optional.of(group));
-		when(groupMemberRepository.existsByUserAndStudyGroup(user2, group)).thenReturn(false);
-
-		//then
-		assertThatThrownBy(() -> studyGroupService.getTopRank(user2, groupId))
-			.isInstanceOf(GroupMemberValidationException.class)
-			.hasFieldOrPropertyWithValue("code", HttpStatus.FORBIDDEN.value())
-			.hasFieldOrPropertyWithValue("error", "랭킹을 확인할 권한이 없습니다.");
-	}
-
-	@Test
 	@DisplayName("스터디 그룹 멤버 역할 수정 성공")
 	void updateGroupMemberRole() {
 		// given
@@ -779,49 +605,5 @@ class StudyGroupServiceTest {
 			.isInstanceOf(GroupMemberValidationException.class)
 			.hasFieldOrPropertyWithValue("code", HttpStatus.BAD_REQUEST.value())
 			.hasFieldOrPropertyWithValue("error", "해당 스터디 그룹에 참여하지 않은 회원입니다.");
-	}
-
-	@Test
-	@DisplayName("랭킹 업데이트 성공")
-	void updateRanking() {
-		// given
-		List<Ranking> rankings = new ArrayList<>();
-		rankings.add(ranking1);
-		rankings.add(ranking2);
-		rankings.add(ranking3);
-		when(rankingRepository.findAllByStudyGroup(group)).thenReturn(rankings);
-		when(rankingRepository.findByMember(groupMember3)).thenReturn(Optional.of(ranking3));
-
-		// when - user3의 랭킹이 +2
-		studyGroupService.updateRanking(groupMember3);
-		studyGroupService.updateRanking(groupMember3);
-		studyGroupService.updateRanking(groupMember3);
-
-		// then
-		assertThat(ranking3.getMember()).isEqualTo(groupMember3);
-		assertThat(ranking3.getCurrentRank()).isEqualTo(1);
-		assertThat(ranking3.getSolvedCount()).isEqualTo(4);
-		assertThat(ranking3.getRankDiff()).isEqualTo("+1");
-
-		assertThat(ranking1.getMember()).isEqualTo(groupMember1);
-		assertThat(ranking1.getCurrentRank()).isEqualTo(2);
-		assertThat(ranking1.getSolvedCount()).isEqualTo(3);
-		assertThat(ranking1.getRankDiff()).isEqualTo("-1");
-
-		assertThat(ranking2.getMember()).isEqualTo(groupMember2);
-		assertThat(ranking2.getCurrentRank()).isEqualTo(3);
-		assertThat(ranking2.getSolvedCount()).isEqualTo(2);
-		assertThat(ranking2.getRankDiff()).isEqualTo("-");
-	}
-
-	@Test
-	@DisplayName("랭킹 업데이트 실패 : 유저 랭킹 정보 조회 실패")
-	void updateRankingFailed_1() {
-		// given
-		when(rankingRepository.findByMember(groupMember3)).thenReturn(Optional.empty());
-		// when, then
-		assertThatThrownBy(() -> studyGroupService.updateRanking(groupMember3))
-			.isInstanceOf(CannotFoundRankingException.class)
-			.hasFieldOrPropertyWithValue("error", "유저의 랭킹 정보를 조회할 수 없습니다.");
 	}
 }
